Moved handle_keyboard_input helper
