--- sim/sim2sim_helpers.py
@@ -1,23 +1,0 @@
-import pygame
-
-
-def handle_keyboard_input():
-    global x_vel_cmd, y_vel_cmd, yaw_vel_cmd
-
-    keys = pygame.key.get_pressed()
-
-    # Update movement commands based on arrow keys
-    if keys[pygame.K_UP]:
-        x_vel_cmd += 0.0005
-    if keys[pygame.K_DOWN]:
-        x_vel_cmd -= 0.0005
-    if keys[pygame.K_LEFT]:
-        y_vel_cmd += 0.0005
-    if keys[pygame.K_RIGHT]:
-        y_vel_cmd -= 0.0005
-
-    # Yaw control
-    if keys[pygame.K_a]:
-        yaw_vel_cmd += 0.001
-    if keys[pygame.K_z]:
-        yaw_vel_cmd -= 0.001
